sync-adapter should cache the own repositories.

+++ mobile/src/main/java/io/syslogic/github/content/QueryStringProvider.java
@@ -7,6 +7,7 @@ import android.database.Cursor;
 import android.net.Uri;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import io.syslogic.github.api.model.QueryString;
 import io.syslogic.github.api.room.Abstraction;
@@ -23,7 +24,9 @@ public class QueryStringProvider extends ContentProvider {
 
     @Override
     public boolean onCreate() {
-        this.dao = Abstraction.getInstance(getContext()).queryStringsDao();
+        if (getContext() != null) {
+            this.dao = Abstraction.getInstance(getContext()).queryStringsDao();
+        }
         return true;
     }
 
@@ -34,27 +37,37 @@ public class QueryStringProvider extends ContentProvider {
 
     @NonNull
     @Override
-    public Cursor query(@NonNull Uri uri, @NonNull String[] projection, @NonNull String selectionClause, @NonNull String[] selectionArgs, @NonNull String sortOrder) {
+    public Cursor query(@NonNull Uri uri, @Nullable String[] projection, @Nullable String selectionClause, @Nullable String[] selectionArgs, @Nullable String sortOrder) {
         return dao.selectAll();
     }
 
     @NonNull
     @Override
-    public Uri insert(@NonNull Uri uri, @NonNull ContentValues values) {
-        QueryString item = new QueryString().fromContentValues(values);
-        long id = dao.insert(item);
-        return ContentUris.withAppendedId(uri, id);
+    public Uri insert(@NonNull Uri uri, @Nullable ContentValues values) {
+        if (values != null) {
+            QueryString item = new QueryString().fromContentValues(values);
+            long id = dao.insert(item);
+            return ContentUris.withAppendedId(uri, id);
+        } else {
+            return uri;
+        }
     }
 
     @Override
-    public int delete(@NonNull Uri uri, @NonNull String selectionClause, @NonNull String[] selectionArgs) {
-        dao.deleteById(Long.valueOf(selectionArgs[0]));
-        return 1;
+    public int delete(@NonNull Uri uri, @Nullable String selectionClause, @Nullable String[] selectionArgs) {
+        if (selectionArgs != null && selectionArgs.length > 0) {
+            return dao.deleteById(Long.valueOf(selectionArgs[0]));
+        } else {
+            return 0;
+        }
     }
 
-    public int update(@NonNull Uri uri, @NonNull ContentValues values, @NonNull String selectionClause, @NonNull String[] selectionArgs) {
-        QueryString item = new QueryString().fromContentValues(values);
-        dao.update(item);
-        return 1;
+    public int update(@NonNull Uri uri, @Nullable ContentValues values, @Nullable String selectionClause, @Nullable String[] selectionArgs) {
+        if (values != null) {
+            QueryString item = new QueryString().fromContentValues(values);
+            return dao.update(item);
+        } else {
+            return 0;
+        }
     }
 }
